# Implement multiple store backends