--- a/cmd/dgraphassigner/main_test.go
+++ b/cmd/dgraphassigner/main_test.go
@@ -21,37 +21,47 @@ func TestQuery(t *testing.T) {
 	minIdx1 := 1 * mod
 
 	logrus.SetLevel(logrus.DebugLevel)
-	dir, err := ioutil.TempDir("", "storetest_")
-	if err != nil {
-		t.Error(err)
-		return
-	}
-	defer os.RemoveAll(dir)
 
-	ps := new(store.Store)
-	ps.Init(dir)
-	clog := commit.NewLogger(dir, "mutations", 50<<20)
-	clog.Init()
-	defer clog.Close()
-	posting.Init(clog)
+	stores := store.Registry.Registered()
 
-	uid.Init(ps)
+	for _, storeName := range stores {
+		t.Log(`Store:`, storeName)
+		dir, err := ioutil.TempDir("", "storetest_")
+		if err != nil {
+			t.Error(err)
+			return
+		}
+		defer os.RemoveAll(dir)
 
-	list := []string{"alice", "bob", "mallory", "ash", "man", "dgraph"}
-	for _, str := range list {
-		if farm.Fingerprint64([]byte(str))%numInstances == 0 {
-			u, err := uid.GetOrAssign(str, 0, numInstances)
-			if u < minIdx0 || u > minIdx0+mod-1 {
-				t.Error("Not the correct UID", err)
+		ps, err := store.Registry.Get(storeName)
+		if err != nil {
+			t.Error(err)
+			return
+		}
+		ps.Init(dir)
+		clog := commit.NewLogger(dir, "mutations", 50<<20)
+		clog.Init()
+		defer clog.Close()
+		posting.Init(clog)
+
+		uid.Init(ps)
+
+		list := []string{"alice", "bob", "mallory", "ash", "man", "dgraph"}
+		for _, str := range list {
+			if farm.Fingerprint64([]byte(str))%numInstances == 0 {
+				u, err := uid.GetOrAssign(str, 0, numInstances)
+				if u < minIdx0 || u > minIdx0+mod-1 {
+					t.Error("Not the correct UID", err)
+				}
+				t.Logf("Instance-0 Correct UID", str, u)
+
+			} else {
+				u, err := uid.GetOrAssign(str, 1, numInstances)
+				if u < minIdx1 || u > minIdx1+mod-1 {
+					t.Error("Not the correct UID", err)
+				}
+				t.Logf("Instance-1 Correct UID", str, u)
 			}
-			t.Logf("Instance-0 Correct UID", str, u)
-
-		} else {
-			u, err := uid.GetOrAssign(str, 1, numInstances)
-			if u < minIdx1 || u > minIdx1+mod-1 {
-				t.Error("Not the correct UID", err)
-			}
-			t.Logf("Instance-1 Correct UID", str, u)
 		}
 	}
 }