--- a/cmd/merge/main.go
+++ b/cmd/merge/main.go
@@ -1,3 +1,5 @@
+// +build rocksdb,cgo
+
 /*
  * Copyright 2015 DGraph Labs, Inc.
  *