The first additional backend is BoltDB.  A `store` flag has been added
to commands, that accepts the name of a store (as of this commit, valid
values are `boltdb` and `rocksdb`).

Stores register their availability via `init()`, which means that new
stores should be imported for side-effects wherever they may be
required directly (predominantly in commands or tests).

Store implementations can be included via build tags.

BREAKING: The default store has been set to `boltdb`, primarily because
this store is native Go, and can be built against even when CGO is not
available.  To produce builds that include rocksdb support, the
`rocksdb` tag _must_ be supplied, ie:

```
go install -tags rocksdb
```

This initial implementation does not include a functional `merge` or `dlist`
command for boltdb.

I saw the comment on #71 regarding the reasoning for choosing RocksDB (it should be noted that while BoltDB only allows a single writer, it does support multiple readers), but the
ability to build without CGO can be preferable for portability, and the important
performance characteristics (read vs write perf for instance) will differ from
project to project, so choice is also desirable.

I have not done significant testing or updated the documentation yet, but discussion would be
welcomed.

The tests/benches should be updated with Go 1.7, as multiple backends are particularly suited to sub-tests/-benchmarks.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/dgraph/127)
<!-- Reviewable:end -->