@manishrjain 2016-07-10
Multiple backend stores are not something we intend to support. There's an expectation of tight coupling between Dgraph and whatever we use to store data on disk. Which we've decided to be RocksDB, given the reasoning that you've right pointed out.

I don't buy the argument that each project is different so read/write concurrency doesn't matter. To begin with, our bulk loader is so efficient purely because of this concurrency that BoltDB doesn't provide. In addition, in live Dgraph instance, we periodically merge dirty posting lists, which currently go unnoticed performance-wise, but would be very negatively impacted if reads and writes become mutually exclusive.

Starting v0.4, which we're aiming to release this week, we'll be making it easier for people to use Dgraph by providing pre-built binaries, that they can readily run on Linux or Mac. That should tackle the main issue you're trying to solve here, which is ease of installation.

Sorry, I can't accept this pull request.