@pdf 2016-07-10
> Multiple backend stores are not something we intend to support.

That could be a valid reason to reject this PR (maintenance complexity, though the surface area is quite small, had you looked at the code), but not for the reasons you've outlined.

> There's an expectation of tight coupling between Dgraph and whatever we use to store data on disk.

Why?  That's certainly not evident from the code except in a couple of the helper utilities.  A KV store is pretty much a KV store, except in edge-cases or performance-related areas.

> Which we've decided to be RocksDB, given the reasoning that you've rightly pointed out.

What reasoning is that?  BoltDB provides good read throughput, so for read-heavy applications, it would possibly be an improvement.

> I don't buy the argument that each project is different so read/write concurrency doesn't matter.

That's not what I said.  I said that projects differ in their requirements for read or write throughput.  For some projects, read throughput is vastly more important than write throughput.

> To begin with, our bulk loader is so efficient purely because of write concurrency, that BoltDB doesn't provide.

I guess that's good, if a project cares about that functionality.

> Also, in live Dgraph instance, we periodically merge dirty posting lists, which currently go unnoticed performance-wise, but would be very negatively impacted if reads and writes become mutually exclusive.

Read operations in BoltDB occur concurrently, writes are serialized.  Do you have any benchmarks to back up your assertions here, or are you just making stuff up?

> If we really wanted to kill concurrency, we could have just used Python.

Sick burn... or something.

> Starting v0.4, which we're aiming to release this week, we'll be making it easier for people to use Dgraph by providing pre-built binaries, which can be readily run on Linux or Mac. That should tackle the main issue you're trying to solve here, which is ease of installation.

Actually, I was looking at embedding Dgraph in a cross-platform/-arch project, and did not want to carry a dependency on a C library.

> Sorry, I can't accept this pull request.

Yeah, looks like Dgraph is not for us.

I have to say that based on the poor attitude in your response here, I think I may be better off not contributing to this project anyway.  I was going to clean up a bunch of other stuff in the code-base after this merge, but I guess you can keep it.