@manishrjain 2016-07-10
I don't think my tone was condescending.

Choosing a library which allows reads and writes concurrently over a library which makes reads and writes mutually exclusive -- is a pretty important design decision, particularly for something that's going to be the basis of Dgraph. I'm not sure what a benchmark is going to prove here, it's basic computer science.

On the other hand, you haven't given any logic for your PR, other than you don't "want" to include any Cgo.