@mohitranka 2016-07-10
@pdf Thanks for your PR and effort. I am sorry that you feel disappointed on the decision of it not getting accepting, which is completely understandable given your effort for this PR.

Whether or not to support multiple backends is an important design decision, and there are successful projects which both ideologies -- eg. Postgres and MySQL. At this point, Dgraph is very young, and have decided to focus on tightly couple backend, as apart from engineering choices (which I am not going to get into, as I am not most qualified when it comes to Go eco-system) it allows to utilise the limited available bandwidth more effectively.

You can find more about the roadmap of dgraph at #1 and discuss major ideas at discuss.dgraph.io.

Once again, we are thankful for your contribution, and sorry for your disappointment. I do however hope you continue to contribute.