@pdf 2016-07-10
@mohitranka thanks, this is a sensible response.  The implementation to support multiple backends is certainly fairly trivial though, as you can see here.

To add some actual data WRT performance, below are the Get/Set benchmarks that were already in the codebase, using 8 cores, for each of the backends.  By default you're using async writes on RocksDB, which may not be safe.  I should also note here that the RocksDB write benchmarks would not complete with less than ~12GB of RAM free, otherwise triggering the OOM killer - something is definitely not right in RocksDB-land (obviously can't use the go memory profiling tools, because cgo).

## RocksDB vs BoltDB (default configuration: RocksDB async, BoltDB sync)
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkGet_valsize1024-8      1473          1669          +13.31%
BenchmarkGet_valsize10KB-8      4468          1678          -62.44%
BenchmarkGet_valsize500KB-8     339189        1680          -99.50%
BenchmarkGet_valsize1MB-8       690710        1682          -99.76%
BenchmarkSet_valsize1024-8      8764          2496036       +28380.56%
BenchmarkSet_valsize10KB-8      116203        2970832       +2456.59%
BenchmarkSet_valsize500KB-8     11092719      12904272      +16.33%
BenchmarkSet_valsize1MB-8       12327737      21963188      +78.16%
```

## RocksDB vs BoltDB (RocksDB async, BoltDB async)
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkGet_valsize1024-8      1473          1662          +12.83%
BenchmarkGet_valsize10KB-8      4468          1750          -60.83%
BenchmarkGet_valsize500KB-8     339189        2110          -99.38%
BenchmarkGet_valsize1MB-8       690710        2211          -99.68%
BenchmarkSet_valsize1024-8      8764          64554         +636.58%
BenchmarkSet_valsize10KB-8      116203        180974        +55.74%
BenchmarkSet_valsize500KB-8     11092719      3853470       -65.26%
BenchmarkSet_valsize1MB-8       12327737      5952093       -51.72%
```

Frequent growth of the BoltDB backing store can hurt performance, but growth size can be increased to reduce this churn:

## RocksDB vs BoltDB (RocksDB async, BoltDB async+AllocSize)
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkGet_valsize1024-8      1473          1654          +12.29%
BenchmarkGet_valsize10KB-8      4468          1641          -63.27%
BenchmarkGet_valsize500KB-8     339189        2167          -99.36%
BenchmarkGet_valsize1MB-8       690710        2195          -99.68%
BenchmarkSet_valsize1024-8      8764          59262         +576.20%
BenchmarkSet_valsize10KB-8      116203        113899        -1.98%
BenchmarkSet_valsize500KB-8     11092719      2856296       -74.25%
BenchmarkSet_valsize1MB-8       12327737      5476866       -55.57%
```

And finally, the safe configuration for both options, using sync writes, without any tuning:

# RocksDB vs BoltDB (RocksDB sync, BoltDB sync)
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkGet_valsize1024-8      1516          1669          +10.09%
BenchmarkGet_valsize10KB-8      4592          1678          -63.46%
BenchmarkGet_valsize500KB-8     353271        1680          -99.52%
BenchmarkGet_valsize1MB-8       680144        1682          -99.75%
BenchmarkSet_valsize1024-8      1825995       2496036       +36.69%
BenchmarkSet_valsize10KB-8      2116578       2970832       +40.36%
BenchmarkSet_valsize500KB-8     6944236       12904272      +85.83%
BenchmarkSet_valsize1MB-8       12376387      21963188      +77.46%
```